chore: consolidate vite/vitest configuration for packages
