+++ src/common/consts/internal/vite.config.ts
@@ -1,3 +1,3 @@
-import { viteConfigFor } from "../../../../infra/build/packages/viteConfigFactory";
+import { viteConfigFor } from "../../../../infra/build/packages/viteConfigFor";
 
 export default viteConfigFor('docodylusConstsInternal')()
